ClemBot.Bot/requirements.txt
aiohttp==3.7.4
aiosqlite==0.12.0
appdirs==1.4.3
astroid==2.3.3
async-timeout==3.0.1
asynctest==0.13.0
attrs==19.3.0
bandit==1.6.2
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
dataclasses-json==0.5.3
dateutils==0.6.8
discord.py==1.6.0
distlib==0.3.1
distro==1.4.0
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
gitdb==4.0.4
GitPython==3.1.1
html5lib==1.0.1
idna==2.9
importlib-metadata==1.7.0
iniconfig==1.0.1
install==1.3.3
ipaddr==2.2.0
isort==4.3.21
lazy-object-proxy==1.4.3
lockfile==0.12.2
mccabe==0.6.1
mock==4.0.2
more-itertools==8.4.0
msgpack==0.6.2
multidict==4.7.5
mypy==0.770
mypy-extensions==0.4.3
numpy==1.20.2
packaging==20.4
pandas==1.2.3
pbr==5.4.5
pep517==0.8.2
Pillow==8.1.1
pluggy==0.13.1
progress==1.5
py==1.10.0
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.4
pyparsing==2.4.7
pyspellchecker==0.5.6
pytest==6.0.1
pytest-asyncio==0.14.0
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.1
PyYAML==5.4
requests==2.24.0
retrying==1.3.3
rope==0.17.0
six==1.14.0
smmap==3.0.2
stevedore==1.32.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==20.0.31
webencodings==0.5.1
websockets==8.1
wrapt==1.11.2
yarl==1.4.2
zipp==3.1.0
